Rewrite/delete some tests that involve v1 tasks.
Replaces test tasks with equivalent rules.

Deletes test_pants_requirement_integration.py, since
it's complicated to port, and the relevant functionality
has sufficient unit tests. In general that was a really
huge integration test just to test the `pants_requirement()`
pseudo-macro, which we may not even need.

Also gets rid of test_backend_independence_integration.py, which
just tested independence of the v1 backend.

[ci skip-rust-tests]
