+++ testprojects/pants-plugins/src/python/test_pants_plugin/BUILD
@@ -1,8 +1,12 @@
+
 python_library(
   dependencies=[
+    'src/python/pants/base:deprecated',
+    'src/python/pants/base:exception_sink',
+    'src/python/pants/engine:goal',
+    'src/python/pants/engine:rules',
+    'src/python/pants/option',
     'testprojects/pants-plugins/3rdparty/python/pants',
-    'testprojects/pants-plugins/src/python/test_pants_plugin/subsystems',
-    'testprojects/pants-plugins/src/python/test_pants_plugin/tasks',
   ],
   provides=setup_py(
     name='test_pants_plugin',
@@ -10,8 +14,6 @@ python_library(
     version='0.0.1',
     namespace_packages=[
       'test_pants_plugin',
-      'test_pants_plugin.subsystems',
-      'test_pants_plugin.tasks',
     ],
   )
 )
