--- testprojects/pants-plugins/src/python/test_pants_plugin/subsystems/BUILD
@@ -1,5 +1,0 @@
-python_library(
-  dependencies=[
-    'testprojects/pants-plugins/3rdparty/python/pants',
-  ],
-)
