--- testprojects/pants-plugins/src/python/test_pants_plugin/tasks/BUILD
@@ -1,9 +1,0 @@
-# Copyright 2018 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_library(
-  dependencies=[
-    'testprojects/pants-plugins/3rdparty/python/pants',
-    'testprojects/pants-plugins/src/python/test_pants_plugin/subsystems',
-  ],
-)
